+++ financial-connections/src/main/java/com/stripe/android/financialconnections/repository/FinancialConnectionsManifestRepository.kt
@@ -7,6 +7,7 @@ import com.stripe.android.core.exception.AuthenticationException
 import com.stripe.android.core.exception.InvalidRequestException
 import com.stripe.android.core.networking.ApiRequest
 import com.stripe.android.financialconnections.analytics.AuthSessionEvent
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsEventContext
 import com.stripe.android.financialconnections.model.AuthorizationRepairResponse
 import com.stripe.android.financialconnections.model.FinancialConnectionsAuthorizationSession
 import com.stripe.android.financialconnections.model.FinancialConnectionsInstitution
@@ -188,7 +189,8 @@ internal interface FinancialConnectionsManifestRepository {
             provideApiRequestOptions: ProvideApiRequestOptions,
             logger: Logger,
             locale: Locale,
-            initialSync: SynchronizeSessionResponse?
+            initialSync: SynchronizeSessionResponse?,
+            eventContext: FinancialConnectionsEventContext
         ): FinancialConnectionsManifestRepository =
             FinancialConnectionsManifestRepositoryImpl(
                 requestExecutor,
@@ -196,7 +198,8 @@ internal interface FinancialConnectionsManifestRepository {
                 provideApiRequestOptions,
                 locale,
                 logger,
-                initialSync
+                initialSync,
+                eventContext
             )
     }
 }
@@ -207,7 +210,8 @@ private class FinancialConnectionsManifestRepositoryImpl(
     val provideApiRequestOptions: ProvideApiRequestOptions,
     val locale: Locale,
     val logger: Logger,
-    initialSync: SynchronizeSessionResponse?
+    initialSync: SynchronizeSessionResponse?,
+    private val eventContext: FinancialConnectionsEventContext
 ) : FinancialConnectionsManifestRepository {
 
     /**
@@ -218,9 +222,14 @@ private class FinancialConnectionsManifestRepositoryImpl(
 
     private val cachedSynchronizeSessionResponseFlow = MutableStateFlow(initialSync)
 
+    init {
+        initialSync?.manifest?.let(eventContext::update)
+    }
+
     private var cachedSynchronizeSessionResponse: SynchronizeSessionResponse?
         get() = cachedSynchronizeSessionResponseFlow.value
         set(value) {
+            value?.manifest?.let(eventContext::update)
             cachedSynchronizeSessionResponseFlow.value = value
         }
 
@@ -260,7 +269,10 @@ private class FinancialConnectionsManifestRepositoryImpl(
             )
         ),
         SynchronizeSessionResponse.serializer()
-    ).also { updateCachedSynchronizeSessionResponse("get/fetch", it) }
+    ).also {
+        require(it.manifest.id.isNotBlank()) { "Financial Connections session ID is missing" }
+        updateCachedSynchronizeSessionResponse("get/fetch", it)
+    }
 
     override suspend fun markConsentAcquired(
         clientSecret: String