+++ financial-connections/src/test/java/com/stripe/android/financialconnections/repository/FinancialConnectionsManifestRepositoryImplTest.kt
@@ -4,6 +4,7 @@ import com.google.common.truth.Truth.assertThat
 import com.stripe.android.core.Logger
 import com.stripe.android.core.networking.ApiRequest
 import com.stripe.android.financialconnections.ApiKeyFixtures
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsEventContext
 import com.stripe.android.financialconnections.domain.GetOrFetchSync.RefetchCondition.None
 import com.stripe.android.financialconnections.model.SynchronizeSessionResponse
 import com.stripe.android.financialconnections.network.FinancialConnectionsRequestExecutor
@@ -30,6 +31,7 @@ internal class FinancialConnectionsManifestRepositoryImplTest {
 
     private val mockRequestExecutor = mock<FinancialConnectionsRequestExecutor>()
     private val apiRequestFactory = mock<ApiRequest.Factory>()
+    private val eventContext = FinancialConnectionsEventContext(null)
 
     private fun buildRepository(
         initialSync: SynchronizeSessionResponse? = null
@@ -43,7 +45,8 @@ internal class FinancialConnectionsManifestRepositoryImplTest {
         },
         logger = Logger.noop(),
         initialSync = initialSync,
-        locale = Locale.US
+        locale = Locale.US,
+        eventContext = eventContext
     )
 
     @Test
@@ -74,6 +77,7 @@ internal class FinancialConnectionsManifestRepositoryImplTest {
             )
 
             verify(mockRequestExecutor, times(1)).execute(any(), any<KSerializer<*>>())
+            assertThat(eventContext.manifest?.id).isEqualTo(ApiKeyFixtures.syncResponse().manifest.id)
         }
 
     @Test
@@ -91,6 +95,7 @@ internal class FinancialConnectionsManifestRepositoryImplTest {
                 )
 
             assertThat(returnedManifest).isEqualTo(initialSync)
+            assertThat(eventContext.manifest).isEqualTo(initialSync.manifest)
             verifyNoInteractions(mockRequestExecutor)
         }